Fixes paginator issue since hugo v0.123.0
Pages of kind "page" don't allow the creation of a paginator (.Paginate
or .Paginator) in the newest version of hugo.
Pages of kind "section" (e.g. /posts) still allow the creation. This fix
adds a check for the kind of page and excludes pages with kind "page"
from creating a paginator.